+++ CHANGELOG.md
@@ -19,9 +19,16 @@
         - do not show progress.
     - `csvtk fix-quotes`:
         - new flag `-b, --buffer-size`.
+    - `csvtk plot`:
+        - new flag `--scale` for scaling the image width/height, tick, axes, line/point and font sizes proportionally, adviced by @tseemann.
     - `csvtk plot line`:
         - only add legend for more than one group. [#279](https://github.com/shenwei356/csvtk/issues/279)
         - sort points by X for plotting lines. [#280](https://github.com/shenwei356/csvtk/issues/280)
+    - `csvtk hist`:
+        - new flags: `--line-width`.
+    - `csvtk box`:
+        - plots of different groups have different colors now.
+        - new flags: `--line-width`, `--point-size`, and `color-index`.
 - [csvtk v0.30.0](https://github.com/shenwei356/csvtk/releases/tag/v0.30.0)
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/csvtk/v0.30.0/total.svg)](https://github.com/shenwei356/csvtk/releases/tag/v0.30.0)
   - `csvtk`: